+++ app/vmalert/web_types.go
@@ -36,13 +36,13 @@ type APIAlert struct {
 
 // WebLink returns a link to the alert which can be used in UI.
 func (aa *APIAlert) WebLink() string {
-	return fmt.Sprintf("alerts?%s=%s&%s=%s",
+	return fmt.Sprintf("alert?%s=%s&%s=%s",
 		paramGroupID, aa.GroupID, paramAlertID, aa.ID)
 }
 
 // APILink returns a link to the alert's JSON representation.
 func (aa *APIAlert) APILink() string {
-	return fmt.Sprintf("api/v1/alerts?%s=%s&%s=%s",
+	return fmt.Sprintf("api/v1/alert?%s=%s&%s=%s",
 		paramGroupID, aa.GroupID, paramAlertID, aa.ID)
 }